requirements.txt
# Core deep learning dependencies
torch>=1.12.0
torch-geometric>=2.1.0
transformers>=4.20.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Chemistry and bioinformatics
rdkit>=2022.03.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.64.0
psutil>=5.8.0

# Optional dependencies for enhanced functionality
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0

# API dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Development and testing (optional)
pytest>=6.0.0
pytest-cov>=2.0.0
black>=22.0.0
flake8>=4.0.0
